feat: move files to `/src` in expo app
