+++ apps/expo/app/_layout.tsx
@@ -1,10 +1,9 @@
 import React from "react";
-import { View } from "react-native";
 import { SafeAreaProvider } from "react-native-safe-area-context";
 import { Stack } from "expo-router";
 import { StatusBar } from "expo-status-bar";
 
-import { TRPCProvider } from "../utils/api";
+import { TRPCProvider } from "../src/utils/api";
 
 // This is the main layout of the app
 // It wraps your pages with the providers they need
