+++ apps/expo/tsconfig.json
@@ -4,5 +4,5 @@
     "strict": true,
     "jsx": "react-native"
   },
-  "include": ["app", "utils", "types", "index.ts", "app.config.ts"]
+  "include": ["src", "app", "index.ts", "app.config.ts"]
 }
