+++ packages/config/tailwind/index.js
@@ -1,12 +1,6 @@
 /** @type {import('tailwindcss').Config} */
 module.exports = {
-  content: [
-    // NextJS
-    "./src/**/*.{ts,tsx}",
-    // Expo
-    "./app/*.{ts,tsx}",
-    "./app/**/*.{ts,tsx}",
-  ],
+  content: ["./src/**/*.{ts,tsx}"],
   theme: {
     extend: {},
   },
